apply naming conventions
