+++ Sources/SQLiteNIOExtras/SQLiteConnectionPool.swift
@@ -1,6 +1,6 @@
 //
 //  SQLiteConnectionPool.swift
-//  feather-sqlite-database
+//  feather-database-sqlite
 //
 //  Created by Tibor Bödecs on 2026. 01. 26..
 //
