+++ Sources/SQLiteNIOExtras/SQLiteConnectionPoolError.swift
@@ -1,6 +1,6 @@
 //
 //  SQLiteConnectionPoolError.swift
-//  feather-sqlite-database
+//  feather-database-sqlite
 //
 //  Created by Tibor Bödecs on 2026. 02. 02..
 //
