+++ Sources/SQLiteNIOExtras/SQLiteTransactionError.swift
@@ -1,6 +1,6 @@
 //
 //  SQLiteTransactionError.swift
-//  feather-sqlite-database
+//  feather-database-sqlite
 //
 //  Created by Tibor Bödecs on 2026. 02. 03..
 //
