+++ Tests/FeatherDatabaseSQLiteTests/FeatherDatabaseSQLiteTestSuite.swift
@@ -1,6 +1,6 @@
 //
-//  FeatherSQLiteDatabaseTestSuite.swift
-//  feather-sqlite-database
+//  FeatherDatabaseSQLiteTestSuite.swift
+//  feather-database-sqlite
 //
 //  Created by Tibor Bödecs on 2026. 01. 10..
 //
@@ -10,7 +10,7 @@ import Logging
 import SQLiteNIO
 import Testing
 
-@testable import FeatherSQLiteDatabase
+@testable import FeatherDatabaseSQLite
 @testable import SQLiteNIOExtras
 
 #if ServiceLifecycleSupport
@@ -18,10 +18,10 @@ import ServiceLifecycleTestKit
 #endif
 
 @Suite
-struct FeatherSQLiteDatabaseTestSuite {
+struct FeatherDatabaseSQLiteTestSuite {
 
     func runUsingTestDatabaseClient(
-        _ closure: ((SQLiteDatabaseClient) async throws -> Void)
+        _ closure: ((DatabaseClientSQLite) async throws -> Void)
     ) async throws {
         var logger = Logger(label: "test")
         logger.logLevel = .info
@@ -33,7 +33,7 @@ struct FeatherSQLiteDatabaseTestSuite {
 
         let client = SQLiteClient(configuration: configuration)
 
-        let database = SQLiteDatabaseClient(
+        let database = DatabaseClientSQLite(
             client: client,
             logger: logger
         )
@@ -1126,12 +1126,151 @@ struct FeatherSQLiteDatabaseTestSuite {
             }
         }
     }
+
+    @Test
+    func rowSequenceIteratesRowsInOrder() async throws {
+        try await runUsingTestDatabaseClient { database in
+            try await database.withConnection { connection in
+                try await connection.run(
+                    query: #"""
+                        CREATE TABLE IF NOT EXISTS "planets" (
+                            "id" INTEGER PRIMARY KEY,
+                            "name" TEXT
+                        );
+                        """#
+                )
+
+                try await connection.run(
+                    query: #"""
+                        INSERT INTO "planets" ("id", "name")
+                        VALUES
+                            (1, 'Mercury'),
+                            (2, 'Venus');
+                        """#
+                )
+
+                let sequence = try await connection.run(
+                    query: #"""
+                        SELECT *
+                        FROM "planets"
+                        ORDER BY "id" ASC;
+                        """#
+                )
+
+                var iterator = sequence.makeAsyncIterator()
+
+                let first = await iterator.next()
+                #expect(first != nil)
+                #expect(
+                    try first?.decode(column: "name", as: String.self)
+                        == "Mercury"
+                )
+
+                let second = await iterator.next()
+                #expect(second != nil)
+                #expect(
+                    try second?.decode(column: "name", as: String.self)
+                        == "Venus"
+                )
+
+                let third = await iterator.next()
+                #expect(third == nil)
+            }
+        }
+    }
+
+    @Test
+    func rowSequenceCollectReturnsAllRows() async throws {
+        try await runUsingTestDatabaseClient { database in
+            try await database.withConnection { connection in
+                try await connection.run(
+                    query: #"""
+                        CREATE TABLE IF NOT EXISTS "greetings" (
+                            "id" INTEGER PRIMARY KEY,
+                            "name" TEXT
+                        );
+                        """#
+                )
+
+                try await connection.run(
+                    query: #"""
+                        INSERT INTO "greetings" ("id", "name")
+                        VALUES
+                            (1, 'Hello'),
+                            (2, 'World');
+                        """#
+                )
+
+                let sequence = try await connection.run(
+                    query: #"""
+                        SELECT
+                            "id",
+                            "name"
+                        FROM "greetings"
+                        ORDER BY "id" ASC;
+                        """#
+                )
+
+                let rows = try await sequence.collect()
+                #expect(rows.count == 2)
+
+                let firstName = try rows[0]
+                    .decode(
+                        column: "name",
+                        as: String.self
+                    )
+                let secondName = try rows[1]
+                    .decode(
+                        column: "name",
+                        as: String.self
+                    )
+
+                #expect(firstName == "Hello")
+                #expect(secondName == "World")
+            }
+        }
+    }
+
+    @Test
+    func rowSequenceHandlesEmptyResults() async throws {
+        try await runUsingTestDatabaseClient { database in
+            try await database.withConnection { connection in
+                try await connection.run(
+                    query: #"""
+                        CREATE TABLE IF NOT EXISTS "empty_rows" (
+                            "id" INTEGER PRIMARY KEY,
+                            "name" TEXT
+                        );
+                        """#
+                )
+
+                let sequence = try await connection.run(
+                    query: #"""
+                        SELECT
+                            "id",
+                            "name"
+                        FROM "empty_rows"
+                        WHERE
+                            1=0;
+                        """#
+                )
+
+                let rows = try await sequence.collect()
+                #expect(rows.isEmpty)
+
+                var iterator = sequence.makeAsyncIterator()
+                let first = await iterator.next()
+                #expect(first == nil)
+            }
+        }
+    }
+
 }
 
 #if ServiceLifecycleSupport
 import ServiceLifecycle
 
-extension FeatherSQLiteDatabaseTestSuite {
+extension FeatherDatabaseSQLiteTestSuite {
 
     @Test
     func serviceLifecycleSupport() async throws {
@@ -1143,8 +1282,8 @@ extension FeatherSQLiteDatabaseTestSuite {
             logger: logger,
         )
         let client = SQLiteClient(configuration: configuration)
-        let database = SQLiteDatabaseClient(client: client, logger: logger)
-        let service = SQLiteDatabaseService(client)
+        let database = DatabaseClientSQLite(client: client, logger: logger)
+        let service = DatabaseServiceSQLite(client)
 
         let serviceGroup = ServiceGroup(
             services: [service],
@@ -1195,7 +1334,7 @@ extension FeatherSQLiteDatabaseTestSuite {
             logger: logger
         )
         let client = SQLiteClient(configuration: configuration)
-        let service = SQLiteDatabaseService(client)
+        let service = DatabaseServiceSQLite(client)
 
         try await withThrowingTaskGroup(of: Void.self) { group in
             group.addTask {
@@ -1232,7 +1371,7 @@ extension FeatherSQLiteDatabaseTestSuite {
             logger: logger
         )
         let client = SQLiteClient(configuration: configuration)
-        let service = SQLiteDatabaseService(client)
+        let service = DatabaseServiceSQLite(client)
         let serviceGroup = ServiceGroup(
             services: [service],
             logger: logger
@@ -1273,7 +1412,7 @@ extension FeatherSQLiteDatabaseTestSuite {
             logger: logger
         )
         let client = SQLiteClient(configuration: configuration)
-        let database = SQLiteDatabaseClient(
+        let database = DatabaseClientSQLite(
             client: client,
             logger: logger
         )
@@ -1308,7 +1447,7 @@ extension FeatherSQLiteDatabaseTestSuite {
             configuration: .init(
                 services: [
                     .init(
-                        service: SQLiteDatabaseService(client)
+                        service: DatabaseServiceSQLite(client)
                     ),
                     .init(
                         service: MigrationService(database: database),
@@ -1342,7 +1481,7 @@ extension FeatherSQLiteDatabaseTestSuite {
             logger: logger
         )
         let client = SQLiteClient(configuration: configuration)
-        let service = SQLiteDatabaseService(client)
+        let service = DatabaseServiceSQLite(client)
 
         try await testGracefulShutdown { trigger in
             try await withThrowingTaskGroup { group in
