+++ Tests/SQLiteNIOExtrasTests/SQLiteNIOExtrasTestSuite.swift
@@ -1,6 +1,6 @@
 //
 //  SQLiteNIOExtrasTestSuite.swift
-//  feather-sqlite-database
+//  feather-database-sqlite
 //
 //  Created by Tibor Bödecs on 2026. 01. 10..
 //
